Refactor ChatViewModel to use selectedModel and add conversation list pinning and tagging tests
- Remove imageModel, ttsModel, and transcriptionModel properties from ChatViewModel.State and replace with selectedModel usage
- Update image generation, transcription, and text-to-speech logic to use selectedModel instead of specific model properties
- Add scrollToBottomTrigger in ChatViewModel.State and trigger scroll on new messages
- Refactor ChatView to move scroll handling into scrollContent method and update to use selectedModel for TTS and image generation UI
- Remove redundant image generation button and related checks in ChatView
- Remove tests for detecting image, transcription, and TTS models and error handling when models are missing from ChatViewModelTests
- Extract pinning-related tests from ConversationListViewModelTests into new ConversationListViewModelTests+Pinning.swift file
- Extract tagging-related tests from ConversationListViewModelTests into new ConversationListViewModelTests+Tags.swift file
- Remove pinning and tagging tests from ConversationListViewModelTests.swift
- Update UserProfileManager to use MainActor.assumeIsolated for iCloud key-value store change handling
- Refactor UserProfileView usage of localized strings in usageSection
- Add new localized strings for user profile, conversation tags, pinning, and personalization features
- Remove obsolete localized error messages related to missing image generation, TTS, and transcription models
