+++ openclient-llm-test/Features/Chat/ChatViewModelTests+Transcription.swift
@@ -13,26 +13,6 @@ import XCTest
 
 @MainActor
 extension ChatViewModelTests {
-    func test_send_viewAppeared_detectsTranscriptionModel() async throws {
-        // Given
-        mockFetchModels.result = .success([
-            LLMModel(id: "gpt-4"),
-            LLMModel(id: "whisper-1", mode: .audioTranscription)
-        ])
-
-        // When
-        sut.send(.viewAppeared)
-        try await Task.sleep(for: .milliseconds(100))
-
-        // Then
-        guard case .loaded(let loadedState) = sut.state else {
-            XCTFail("Expected loaded state")
-            return
-        }
-        XCTAssertEqual(loadedState.transcriptionModel?.id, "whisper-1")
-        XCTAssertEqual(loadedState.availableModels.count, 1, "Should not include the transcription model")
-    }
-
     func test_send_audioRecorded_setsIsTranscribing() async throws {
         // Given
         let mockTranscribe = MockTranscribeAudioUseCase()
@@ -163,34 +143,4 @@ extension ChatViewModelTests {
         XCTAssertEqual(loadedState.inputText, "")
     }
 
-    func test_send_audioRecorded_withoutTranscriptionModel_showsError() async throws {
-        // Given
-        let mockTranscribe = MockTranscribeAudioUseCase()
-        mockFetchModels.result = .success([LLMModel(id: "gpt-4")])
-
-        sut = ChatViewModel(
-            fetchModelsUseCase: mockFetchModels,
-            streamMessageUseCase: mockStreamMessage,
-            saveConversationUseCase: mockSaveConversation,
-            transcribeAudioUseCase: mockTranscribe,
-            settingsManager: mockSettingsManager,
-            conversationStartersManager: mockConversationStarters
-        )
-
-        sut.send(.viewAppeared)
-        try await Task.sleep(for: .milliseconds(100))
-
-        // When
-        sut.send(.audioRecorded(Data([1, 2, 3]), 2.0))
-        try await Task.sleep(for: .milliseconds(100))
-
-        // Then
-        guard case .loaded(let loadedState) = sut.state else {
-            XCTFail("Expected loaded state")
-            return
-        }
-        XCTAssertFalse(mockTranscribe.executeCalled)
-        XCTAssertNotNil(loadedState.errorMessage)
-        XCTAssertFalse(loadedState.isTranscribing)
-    }
 }
