+++ openclient-llm-test/Features/Chat/ConversationListViewModelTests.swift
@@ -402,157 +402,4 @@ final class ConversationListViewModelTests: XCTestCase {
         }
         XCTAssertTrue(loadedState.filteredConversations.isEmpty)
     }
-
-    // MARK: - Tests — pinToggled
-
-    func test_send_pinToggled_pinsConversation() async throws {
-        // Given
-        let conversation = Conversation(modelId: "gpt-4")
-        mockLoadConversations.result = .success([conversation])
-        mockFetchModels.result = .success([])
-        sut.send(.viewAppeared)
-        try await Task.sleep(for: .milliseconds(100))
-
-        // When
-        sut.send(.pinToggled(conversation.id))
-
-        // Then
-        guard case .loaded(let loadedState) = sut.state else {
-            XCTFail("Expected loaded state")
-            return
-        }
-        XCTAssertEqual(mockPinConversation.executedId, conversation.id)
-        XCTAssertEqual(mockPinConversation.executedIsPinned, true)
-        XCTAssertTrue(loadedState.conversations.first?.isPinned ?? false)
-    }
-
-    func test_send_pinToggled_unpinsAlreadyPinnedConversation() async throws {
-        // Given
-        let conversation = Conversation(modelId: "gpt-4", isPinned: true)
-        mockLoadConversations.result = .success([conversation])
-        mockFetchModels.result = .success([])
-        sut.send(.viewAppeared)
-        try await Task.sleep(for: .milliseconds(100))
-
-        // When
-        sut.send(.pinToggled(conversation.id))
-
-        // Then
-        XCTAssertEqual(mockPinConversation.executedIsPinned, false)
-
-        guard case .loaded(let loadedState) = sut.state else {
-            XCTFail("Expected loaded state")
-            return
-        }
-        XCTAssertFalse(loadedState.conversations.first?.isPinned ?? true)
-    }
-
-    func test_send_pinToggled_pinnedConversationsAppearInPinnedSection() async throws {
-        // Given
-        let pinned = Conversation(title: "Pinned", modelId: "gpt-4", isPinned: true)
-        let unpinned = Conversation(title: "Unpinned", modelId: "gpt-4")
-        mockLoadConversations.result = .success([pinned, unpinned])
-        mockFetchModels.result = .success([])
-        sut.send(.viewAppeared)
-        try await Task.sleep(for: .milliseconds(100))
-
-        // Then
-        guard case .loaded(let loadedState) = sut.state else {
-            XCTFail("Expected loaded state")
-            return
-        }
-        let pinnedSection = loadedState.groupedConversations.first(where: { $0.period == .pinned })
-        XCTAssertNotNil(pinnedSection)
-        XCTAssertEqual(pinnedSection?.conversations.count, 1)
-        XCTAssertEqual(pinnedSection?.conversations.first?.title, "Pinned")
-    }
-
-    // MARK: - Tests — tagsUpdated
-
-    func test_send_tagsUpdated_savesTagsForConversation() async throws {
-        // Given
-        let conversation = Conversation(modelId: "gpt-4")
-        mockLoadConversations.result = .success([conversation])
-        mockFetchModels.result = .success([])
-        sut.send(.viewAppeared)
-        try await Task.sleep(for: .milliseconds(100))
-
-        // When
-        sut.send(.tagsUpdated(conversation.id, ["swift", "ai"]))
-
-        // Then
-        XCTAssertEqual(mockUpdateTags.executedId, conversation.id)
-        XCTAssertEqual(mockUpdateTags.executedTags, ["swift", "ai"])
-
-        guard case .loaded(let loadedState) = sut.state else {
-            XCTFail("Expected loaded state")
-            return
-        }
-        XCTAssertEqual(loadedState.conversations.first?.tags, ["swift", "ai"])
-    }
-
-    func test_send_tagsUpdated_allTagsComputedCorrectly() async throws {
-        // Given
-        let c1 = Conversation(modelId: "gpt-4", tags: ["swift", "ai"])
-        let c2 = Conversation(modelId: "gpt-4", tags: ["ai", "coding"])
-        mockLoadConversations.result = .success([c1, c2])
-        mockFetchModels.result = .success([])
-        sut.send(.viewAppeared)
-        try await Task.sleep(for: .milliseconds(100))
-
-        // Then
-        guard case .loaded(let loadedState) = sut.state else {
-            XCTFail("Expected loaded state")
-            return
-        }
-        // allTags should be sorted and deduplicated
-        XCTAssertEqual(loadedState.allTags, ["ai", "coding", "swift"])
-    }
-
-    // MARK: - Tests — tagFilterChanged
-
-    func test_send_tagFilterChanged_filtersConversationsByTag() async throws {
-        // Given
-        let c1 = Conversation(title: "Swift Chat", modelId: "gpt-4", tags: ["swift"])
-        let c2 = Conversation(title: "AI Chat", modelId: "gpt-4", tags: ["ai"])
-        mockLoadConversations.result = .success([c1, c2])
-        mockFetchModels.result = .success([])
-        sut.send(.viewAppeared)
-        try await Task.sleep(for: .milliseconds(100))
-
-        // When
-        sut.send(.tagFilterChanged("swift"))
-
-        // Then
-        guard case .loaded(let loadedState) = sut.state else {
-            XCTFail("Expected loaded state")
-            return
-        }
-        XCTAssertEqual(loadedState.filteredConversations.count, 1)
-        XCTAssertEqual(loadedState.filteredConversations.first?.title, "Swift Chat")
-        XCTAssertEqual(loadedState.activeTagFilter, "swift")
-    }
-
-    func test_send_tagFilterChanged_nil_showsAll() async throws {
-        // Given
-        let c1 = Conversation(title: "Swift Chat", modelId: "gpt-4", tags: ["swift"])
-        let c2 = Conversation(title: "AI Chat", modelId: "gpt-4", tags: ["ai"])
-        mockLoadConversations.result = .success([c1, c2])
-        mockFetchModels.result = .success([])
-        sut.send(.viewAppeared)
-        try await Task.sleep(for: .milliseconds(100))
-
-        sut.send(.tagFilterChanged("swift"))
-
-        // When
-        sut.send(.tagFilterChanged(nil))
-
-        // Then
-        guard case .loaded(let loadedState) = sut.state else {
-            XCTFail("Expected loaded state")
-            return
-        }
-        XCTAssertEqual(loadedState.filteredConversations.count, 2)
-        XCTAssertNil(loadedState.activeTagFilter)
-    }
 }
