+++ openclient-llm/Shared/Resources/Localizable.xcstrings
@@ -21,6 +21,22 @@
     "%lld tokens" : {
 
     },
+    "%lld/%lld" : {
+      "comment" : "A label showing the current character count and the maximum allowed.",
+      "isCommentAutoGenerated" : true,
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "new",
+            "value" : "%1$lld/%2$lld"
+          }
+        }
+      }
+    },
+    "A brief description about yourself. Max 200 characters." : {
+      "comment" : "A description of a user's profile description.",
+      "isCommentAutoGenerated" : true
+    },
     "A network error occurred. Please try again." : {
 
     },
@@ -30,8 +46,16 @@
     "Accepted" : {
 
     },
+    "Add" : {
+      "comment" : "A button that adds a tag.",
+      "isCommentAutoGenerated" : true
+    },
     "Add a comment" : {
 
+    },
+    "Add tag..." : {
+      "comment" : "A placeholder for a text field that adds a tag to a conversation.",
+      "isCommentAutoGenerated" : true
     },
     "All" : {
 
@@ -44,6 +68,10 @@
     },
     "Answer a tricky question" : {
 
+    },
+    "Any additional context you want the assistant to know. Max 500 characters." : {
+      "comment" : "A description of the extra information section.",
+      "isCommentAutoGenerated" : true
     },
     "API Key (Optional)" : {
 
@@ -115,6 +143,10 @@
     "Completed" : {
 
     },
+    "Configure your name and personal context to personalise model responses." : {
+      "comment" : "A description of the personalization section.",
+      "isCommentAutoGenerated" : true
+    },
     "Connection successful" : {
 
     },
@@ -157,6 +189,10 @@
     "Describe your suggestion in detail..." : {
 
     },
+    "Description" : {
+      "comment" : "A label displayed above the user's profile description.",
+      "isCommentAutoGenerated" : true
+    },
     "Description (optional)" : {
 
     },
@@ -171,6 +207,10 @@
     },
     "Earlier" : {
       "comment" : "Title for a section of conversation data that includes conversations older than a week.",
+      "isCommentAutoGenerated" : true
+    },
+    "Edit Tags" : {
+      "comment" : "A button that opens a sheet for editing a conversation's tags.",
       "isCommentAutoGenerated" : true
     },
     "Enter a brief title for the issue" : {
@@ -193,6 +233,10 @@
     },
     "Explain why this feature would be useful" : {
 
+    },
+    "Extra Info" : {
+      "comment" : "A label displayed above the user's extra information.",
+      "isCommentAutoGenerated" : true
     },
     "Feedback" : {
 
@@ -221,6 +265,10 @@
     },
     "How can I help you?" : {
 
+    },
+    "How the assistant will address you. Max 50 characters." : {
+      "comment" : "A description of how the assistant will address the user.",
+      "isCommentAutoGenerated" : true
     },
     "iCloud Sync" : {
 
@@ -285,6 +333,10 @@
     "Models" : {
 
     },
+    "Name" : {
+      "comment" : "A label displayed above the user's name.",
+      "isCommentAutoGenerated" : true
+    },
     "New Chat" : {
 
     },
@@ -293,6 +345,10 @@
     },
     "New suggestion" : {
 
+    },
+    "New Tag" : {
+      "comment" : "A label displayed above a text field to add a new tag.",
+      "isCommentAutoGenerated" : true
     },
     "Next" : {
 
@@ -306,10 +362,6 @@
     "No Conversations" : {
 
     },
-    "No image generation models available. Add a model like DALL·E or gpt-image-1 to your LiteLLM server." : {
-      "comment" : "Error message displayed when no image generation models are available.",
-      "isCommentAutoGenerated" : true
-    },
     "No internet connection. Please check your network." : {
 
     },
@@ -318,14 +370,6 @@
     },
     "No suggestions yet." : {
 
-    },
-    "No text-to-speech models available. Add a TTS model like tts-1 to your LiteLLM server." : {
-      "comment" : "Error message displayed when there are no text-to-speech models available.",
-      "isCommentAutoGenerated" : true
-    },
-    "No transcription models available. Add a model like whisper-1 to your LiteLLM server." : {
-      "comment" : "Error message displayed when there are no available transcription models.",
-      "isCommentAutoGenerated" : true
     },
     "Nucleus sampling. Lower values make output more focused." : {
 
@@ -359,8 +403,24 @@
     "Pending" : {
 
     },
+    "Personal Context" : {
+      "comment" : "A button that opens a sheet for configuring the user's name and personal context.",
+      "isCommentAutoGenerated" : true
+    },
+    "Personalization" : {
+      "comment" : "A heading for the personalization settings.",
+      "isCommentAutoGenerated" : true
+    },
     "Photo Library" : {
 
+    },
+    "Pin" : {
+      "comment" : "A pin icon.",
+      "isCommentAutoGenerated" : true
+    },
+    "Pinned" : {
+      "comment" : "Title for the section of conversations that are pinned.",
+      "isCommentAutoGenerated" : true
     },
     "Post" : {
 
@@ -499,6 +559,10 @@
     "Summarize a long text" : {
 
     },
+    "Swipe left to remove a tag." : {
+      "comment" : "A footer displayed under the list of tags.",
+      "isCommentAutoGenerated" : true
+    },
     "Sync" : {
       "comment" : "A heading for the sync settings.",
       "isCommentAutoGenerated" : true
@@ -509,6 +573,10 @@
     },
     "System Prompt" : {
 
+    },
+    "Tags" : {
+      "comment" : "A heading displayed above the user's tags.",
+      "isCommentAutoGenerated" : true
     },
     "Tap + to get started" : {
 
@@ -549,6 +617,10 @@
     "Thinking..." : {
 
     },
+    "This information is added to every conversation so models can personalise their responses." : {
+      "comment" : "A description of the information that is added to every conversation.",
+      "isCommentAutoGenerated" : true
+    },
     "This Week" : {
       "comment" : "Title of a conversation section for conversations from the current week.",
       "isCommentAutoGenerated" : true
@@ -574,6 +646,10 @@
     },
     "Unable to Load Models" : {
 
+    },
+    "Unpin" : {
+      "comment" : "A label for un-pinning a conversation.",
+      "isCommentAutoGenerated" : true
     },
     "Very creative" : {
 
@@ -601,6 +677,10 @@
     },
     "Your gateway to any LLM through a unified interface. Connect to your LiteLLM server and start chatting." : {
 
+    },
+    "Your name" : {
+      "comment" : "A label that describes the user's name.",
+      "isCommentAutoGenerated" : true
     },
     "Your name (optional)" : {
